Move view switching commands to editor and metadata packages.
