+++ src/article/shared/SwitchViewCommand.js
